imagenet/requirements.txt
torch
pillow
